feat: check forbidden for DeletePostUseCase

+++ src/features/blog/application/useCases/delete.post.by.blog.id.use.case.ts
@@ -2,9 +2,14 @@ import { CommandHandler } from '@nestjs/cqrs';
 import { PostRepositoryTypeorm } from '@post/repository/typeorm/post.repository.typeorm';
 import { PostQueryRepositoryTypeorm } from '@post/repository/typeorm/post.query.repository.typeorm';
 import { BlogQueryRepositoryTypeorm } from '@blog/repository/typeorm/blog.query.repository.typeorm';
+import { ForbiddenException } from '@nestjs/common';
 
 export class DeletePostByBlogIdCommand {
-  constructor(public postId: string, public blogId: string) {}
+  constructor(
+    public postId: string,
+    public blogId: string,
+    public userId: string
+  ) { }
 }
 
 @CommandHandler(DeletePostByBlogIdCommand)
@@ -13,14 +18,15 @@ export class DeletePostByBlogIdUseCase {
     private blogQueryRepository: BlogQueryRepositoryTypeorm,
     private postRepository: PostRepositoryTypeorm,
     private postQueryrepository: PostQueryRepositoryTypeorm,
-  ) {}
+  ) { }
 
   async execute(command: DeletePostByBlogIdCommand): Promise<boolean> {
-    const { postId, blogId } = command;
+    const { postId, blogId, userId } = command;
 
-    const blog = await this.blogQueryRepository.findBlogViewById(blogId);
+    const blog = await this.blogQueryRepository.findFullBlogById(blogId);
     const post = await this.postQueryrepository.findFullPostById(postId);
     if (!blog || !post) return false;
+    if (blog.user.id !== userId) throw new ForbiddenException()
 
     await this.postRepository.deletePostById(post.id);
     return true;
